+++ lib/deimos/test_helpers.rb
@@ -30,6 +30,14 @@ module Deimos
         end
       end
 
+      # Kafka test config with avro schema registry
+      def kafka_schema_registry_test!
+        Deimos.configure do |deimos_config|
+          deimos_config.producers.backend = :kafka
+          deimos_config.schema.backend = :avro_schema_registry
+        end
+      end
+
       # Set the config to the right settings for a kafka test
       def kafka_test!
         Deimos.configure do |deimos_config|